+++ tests/webfinger/client/4_3__6_accept_example_response.py
@@ -1,11 +1,11 @@
 import json
 from hamcrest import assert_that
 
-from feditest import step
+from feditest import test
 from feditest.protocols.webfinger import WebFingerClient, WebFingerServer
 from feditest.protocols.webfinger.traffic import WebFingerQueryResponse
 
-@step
+@test
 def accept_example_response(
         client: WebFingerClient,
         server: WebFingerServer
