+++ tests/webfinger/client/4_4_1__3_accept_jrds_without_subject.py
@@ -1,11 +1,11 @@
 import json
 from hamcrest import assert_that
 
-from feditest import step
+from feditest import test
 from feditest.protocols.webfinger import WebFingerClient, WebFingerServer
 from feditest.protocols.webfinger.traffic import WebFingerQueryResponse
 
-@step
+@test
 def accept_jrds_without_subject(
         client: WebFingerClient,
         server: WebFingerServer
