+++ tests/webfinger/server/4_1__2_parameter_ordering_not_significant.py
@@ -1,7 +1,8 @@
-from feditest import step
+from feditest import test
 from feditest.protocols.webfinger import WebFingerClient, WebFingerServer
 
-@step
+
+@test
 def parameter_ordering(
         client: WebFingerClient,
         server: WebFingerServer
